requirements.txt
torch
torchvision
pillow
requests
pynacl
